requirements.txt
numpy==1.24
pytorch==1.13
scipy==1.11
jpype1==1.5.0